Add favicon and rename views dir to pages
