+++ src/index.js
@@ -1,7 +1,7 @@
 import { render } from "lit-html";
 import './styles/base.css'
 
-import PlayerDetails from './views/player-details'
+import PlayerDetails from './pages/PlayerDetails'
 
 document.addEventListener("DOMContentLoaded", () => {
   render(PlayerDetails, document.body);
