+++ src/pages/PlayerDetails/index.js
@@ -1,92 +1,7 @@
 import { html } from "lit-html";
-import { css } from "linaria";
 
-import { keyboardIcon, psIcon, xboxIcon } from "../svg";
-
-const container = css`
-  min-height: 100vh;
-  padding: 6.25rem 1rem 3.125rem;
-
-  @media (min-width: 600px) {
-    width: 500px;
-    margin: 0 auto;
-  }
-`;
-
-const logo = css`
-  display: block;
-  width: 14.25rem;
-  height: 13rem;
-  margin: 0 auto 2rem;
-
-  svg {
-    width: 100%;
-    height: 100%;
-
-    g {
-      fill: var(--color-primary);
-    }
-  }
-`;
-
-const platformChoice = css`
-  display: flex;
-  justify-content: space-between;
-
-  div {
-    flex: 1;
-    display: flex;
-    justify-content: center;
-    padding: 0.5rem;
-
-    &.selected svg {
-      path,
-      g {
-        fill: var(--color-primary);
-      }
-    }
-  }
-
-  input {
-    display: none;
-  }
-
-  label {
-    display: block;
-    width: 100%;
-    height: 4rem;
-    cursor: pointer;
-  }
-
-  svg {
-    display: block;
-    height: 100%;
-    margin: 0 auto;
-  }
-`;
-
-const formGroup = css`
-  padding: 2rem 0 0;
-  display: flex;
-  flex-direction: column;
-
-  input {
-    padding: 1rem 0.5rem;
-    border-width: 0 0 2px;
-    border-color: var(--color-primary);
-    font-size: 1.25rem;
-  }
-`;
-
-const confirmBtn = css`
-  margin: 3rem 0 0;
-  width: 100%;
-  height: 4.6875rem;
-  background-color: var(--color-primary);
-  border-radius: 3px;
-  font-size: 2rem;
-  color: var(--color-magnolia);
-`;
+import * as styles from './styles'
+import { keyboardIcon, psIcon, xboxIcon } from "../../svg";
 
 const onConfirmPlayerDetails = () => {
   const platformChoice = document.querySelector("#platform").value;
@@ -96,8 +11,8 @@ const onConfirmPlayerDetails = () => {
 };
 
 const markup = html`
-  <section class=${container}>
-    <div class=${logo}>
+  <section class=${styles.container}>
+    <div class=${styles.logo}>
       <svg xmlns="http://www.w3.org/2000/svg">
         <g fill="#000" fill-rule="nonzero">
           <path
@@ -109,9 +24,9 @@ const markup = html`
     <h1>Legend Alert</h1>
     <h2>Enter your player details below</h2>
     <form>
-      <div class=${formGroup}>
+      <div class=${styles.formGroup}>
         <h3>Which Platform?</h3>
-        <div class=${platformChoice}>
+        <div class=${styles.platformChoice}>
           <div>
             <input type="radio" id="platform" name="platform" value="psn" />
             <label for="psn">${psIcon}</label>
@@ -126,7 +41,7 @@ const markup = html`
           </div>
         </div>
       </div>
-      <div class=${formGroup}>
+      <div class=${styles.formGroup}>
         <h3>Player tag</h3>
         <input
           type="text"
@@ -136,7 +51,7 @@ const markup = html`
         />
       </div>
       <button
-        class=${confirmBtn}
+        class=${styles.confirmBtn}
         @click=${onConfirmPlayerDetails}
         type="button"
       >
