+++ src/static/index.html
@@ -2,6 +2,10 @@
 <html>
 <head>
   <link href="https://fonts.googleapis.com/css?family=Nova+Square&display=swap" rel="stylesheet">
+  <link rel="apple-touch-icon" sizes="180x180" href="/images/apple-touch-icon.png">
+  <link rel="icon" type="image/png" sizes="32x32" href="/images/favicon-32x32.png">
+  <link rel="icon" type="image/png" sizes="16x16" href="/images/favicon-16x16.png">
+  <!-- <link rel="manifest" href="/site.webmanifest"> -->
 </head>
 <body></body>
 </html>
